Stop using allocation trick in fastpath functions (OCaml 4.02).
In OCaml 4.02, there is a new compiler optimization which means we can
no longer rely on the trick of passing in a newly allocated `zero'
(ie. Int32.zero or Int64.zero) to our fastpath noalloc functions.

Instead we have to turn these into regular functions, which means they
may be a bit slower.