configure.ac: make the `time` command optional
The `time` command is only required for the `benchmarks` Makefile
target which is only used for development, not for installing the
package. It is unfortunate to block installation on systems that could
perfectly support it (the problem is compounded by the fact that the
target uses GNU-specific time options that are not available on the
`time` command of BSD or OSX systems).

This change makes the configure only warn, instead of erroring, if
`time` is missing. I checked that the rest of the build goes perfectly
well -- except the `benchmarks` rule is now a no-op.